text2vid/pyproject.toml
[tool.poetry]
name = "text2vid"
version = "0.1.0"
description = ""
authors = ["Sway Ducky <swayducky@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
accelerate = "0.17.1"
huggingface-hub = "0.13.3"
imageio = {version = "2.26.1", extras = ["ffmpeg"]}
torch = "1.13.0"
transformers = "4.27.2"
diffusers = {path = "diffusers"}
python-slugify = "^8.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
